Refactoring: Move interceptors to frontend
