+++ pkg/auth/azure.go
@@ -28,10 +28,6 @@ import (
 
 	"github.com/MicahParks/keyfunc/v2"
 	jwt "github.com/golang-jwt/jwt/v5"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/metadata"
-	"google.golang.org/grpc/status"
 )
 
 func JWKSInitAzureFromJson() (*keyfunc.JWKS, error) {
@@ -99,69 +95,6 @@ func ValidateToken(jwtB64 string, jwks *keyfunc.JWKS, clientId string, tenantId 
 	}
 
 	return claims, nil
-}
-
-func authorize(ctx context.Context, jwks *keyfunc.JWKS, clientId string, tenantId string) (*User, error) {
-	md, ok := metadata.FromIncomingContext(ctx)
-	if !ok {
-		return nil, status.Errorf(codes.InvalidArgument, "Retrieving metadata failed")
-	}
-
-	authHeader, ok := md["authorization"]
-	if !ok {
-		return nil, status.Errorf(codes.Unauthenticated, "Authorization token not supplied")
-	}
-
-	token := authHeader[0]
-	claims, err := ValidateToken(token, jwks, clientId, tenantId)
-	if err != nil {
-		return nil, status.Errorf(codes.Unauthenticated, "Invalid authorization token provided")
-	}
-	u := MakeDefaultUser()
-	if _, ok := claims["aud"]; ok && claims["aud"] == clientId {
-		u = &User{
-			Email: claims["email"].(string),
-			Name:  claims["name"].(string),
-		}
-	}
-
-	return u, nil
-}
-
-func UnaryInterceptor(ctx context.Context,
-	req interface{},
-	info *grpc.UnaryServerInfo,
-	handler grpc.UnaryHandler,
-	jwks *keyfunc.JWKS,
-	clientId string,
-	tenantId string) (interface{}, error) {
-	if info.FullMethod != "/api.v1.FrontendConfigService/GetConfig" {
-		userData, err := authorize(ctx, jwks, clientId, tenantId)
-		if err != nil {
-			return nil, err
-		}
-		ctx = ToContext(ctx, userData)
-	}
-	h, err := handler(ctx, req)
-	return h, err
-
-}
-
-func StreamInterceptor(
-	srv interface{},
-	stream grpc.ServerStream,
-	info *grpc.StreamServerInfo,
-	handler grpc.StreamHandler,
-	jwks *keyfunc.JWKS,
-	clientId string,
-	tenantId string,
-
-) error {
-	_, err := authorize(stream.Context(), jwks, clientId, tenantId)
-	if err != nil {
-		return err
-	}
-	return handler(srv, stream)
 }
 
 func HttpAuthMiddleWare(resp http.ResponseWriter, req *http.Request, jwks *keyfunc.JWKS, clientId string, tenantId string, allowedPaths []string, allowedPrefixes []string) error {
